New wi::bitreverse function.
This patch provides a wide-int implementation of bitreverse, that
implements both of Richard Sandiford's suggestions from the review at
https://gcc.gnu.org/pipermail/gcc-patches/2023-May/618215.html of an
improved API (as a stand-alone function matching the bswap refactoring),
and an implementation that works with any bit-width precision.

This patch has been tested on x86_64-pc-linux-gnu with make bootstrap
(and a make check-gcc).  Ok for mainline?  Are the remaining pieces
of the above patch pre-approved (pending re-testing)?  The aim is that
this new code will be thoroughly tested by the new *-2.c test cases in
https://gcc.gnu.org/git/?p=gcc.git;h=c09471fbc7588db2480f036aa56a2403d3c03ae
5
with a minor tweak to use the BITREVERSE rtx in the NVPTX back-end,
followed by similar tests on other targets that provide bit-reverse
built-ins (such as ARM and xstormy16), in advance of support for a
backend-independent solution to PR middle-end/50481.

2023-06-02  Roger Sayle  <roger@nextmovesoftware.com>

gcc/ChangeLog
        * wide-int.cc (wi::bitreverse_large): New function implementing
        bit reversal of an integer.
        * wide-int.h (wi::bitreverse): New (template) function prototype.
        (bitreverse_large): Prototype helper function/implementation.
        (wi::bitreverse): New template wrapper around bitreverse_large.

Thanks again,
Roger