c++: is_specialization_of_friend confusion [PR109923]
Bootstrapped and regtested on x86_64-pc-linux-gnu, does this look OK for
trunk?

-- >8 --

The check for a non-template member function of a class template in
is_specialization_of_friend is overbroad, and accidentally holds for a
non-template hidden friend too, which causes the predicate to return
true for

  decl = void non_templ_friend(A<int>, A<void>)
  friend_decl = void non_templ_friend(A<void>, A<void>)

This patch refines the check appropriately.

	PR c++/109923

gcc/cp/ChangeLog:

	* pt.cc (is_specialization_of_friend): Fix overbroad check for
	a non-template member function of a class template.

gcc/testsuite/ChangeLog:

	* g++.dg/template/friend79.C: New test.