c++: bad 'this' conversion for nullary memfn [PR106760]
Bootstrapped and regtested on x86_64-pc-linu-xgnu, does this look OK for trunk?

-- >8 --

Here we notice the 'this' conversion for the call f<void>() is bad, so
we correctly defer instantiating it, but we end up never adding it to
'bad_fns' since missing_conversion_p for it returns false (its only
argument is the already computed 'this' argument).  This is not a huge
deal, but it causes us to longer accept the call with -fpermissive.

So if we have a non-strictly viable template candidate that has not been
instantiated, then we need to add it to 'bad_fns' even if it doesn't
have any missing conversions.

	PR c++/106760

gcc/cp/ChangeLog:

	* call.cc (add_candidates): Relax test for adding a candidate
	to 'bad_fns' to accept an uninstantiated template candidate that
	has no missing conversions.

gcc/testsuite/ChangeLog:

	* g++.dg/ext/conv3.C: New test.