Convert H8 port to LRA
With Vlad's recent LRA fix to the elimination code, the H8 can be
converted to LRA.

This patch has two changes of note.

First, this turns Zz into a standard constraint.  This helps reloading
for the H8/SX movqi pattern.

Second, this drops the whole pattern for the SX bit memory operations.
I can't see why those exist to begin with.  They should be handled by
the standard bit manipulation patterns.   If someone wants to try and
improve SX bit support, that'd be great and they can do so within the
LRA framework :-)

Pushed to the trunk...

Jeff
commit f66e0a94ad7bc18538c8207fc2c86b62e4a51bb2
Author: Jeff Law <jlaw@ventanamicro.com>
Date:   Sun Jun 4 11:38:55 2023 -0600

    Convert H8 port to LRA

    With Vlad's recent LRA fix to the elimination code, the H8 can be converted
    to LRA.

    This patch has two changes of note.

    First, this turns Zz into a standard constraint.  This helps reloading for
    the H8/SX movqi pattern.

    Second, this drops the whole pattern for the SX bit memory operations.  I
    can't see why those exist to begin with.  They should be handled by the
    standard bit manipulation patterns.   If someone wants to try and improve SX
    bit support, that'd be great and they can do so within the LRA framework :-)

    Pushed to the trunk...

    gcc/
            * config/h8300/constraints.md (Zz): Make this a normal
            constraint.
            * config/h8300/h8300.cc (TARGET_LRA_P): Remove.
            * config/h8300/logical.md (H8/SX bit patterns): Remove.