libstdc++: Avoid vector casts while still avoiding PR90424
This is the first part of a fix for the failure of the new simd test on non-
VSX POWER targets. There are more casts to unavailable vectors of 64-bit
element types to be rewritten.

OK for master and backports?

Tested on x86_64-pc-linux-gnu and powerpc64le-linux-gnu

------ >8 -------

Signed-off-by: Matthias Kretz <m.kretz@gsi.de>

libstdc++-v3/ChangeLog:

	PR libstdc++/109822
	* include/experimental/bits/simd_builtin.h (_S_store): Rewrite
	to avoid casts to other vector types. Implement store as
	succession of power-of-2 sized memcpy to avoid PR90424.