c++: NRV and goto [PR92407]
Tested x86_64-pc-linux-gnu, applying to trunk.

-- 8< --

Here our named return value optimization was breaking the required
destructor when the goto takes 'a' out of scope.  The simplest fix is to
disable the optimization in the presence of user labels.

We could do better by disabling the optimization only if there is a backward
goto across the variable declaration, but we don't currently track that.

	PR c++/92407

gcc/cp/ChangeLog:

	* typeck.cc (check_return_expr): Prevent NRV in the presence of
	named labels.

gcc/testsuite/ChangeLog:

	* g++.dg/opt/nrv22.C: New test.