LoongArch:Change the default value of LARCH_CALL_RATIO to 6 on the LoongArch architecture.
During the regression testing of the LoongArch architecture GCC, it was
found that the tests in the pr90883.C file failed. The problem was modulated and
found that the error was caused by setting the macro LARCH_CALL_RATIO to a too
large value. Combined with the actual LoongArch architecture, the different
thresholds for meeting the test conditions were tested using the engineering method
(SPEC CPU 2006), and the results showed that its optimal threshold should be set
to 6.
gcc/ChangeLog:

	* config/loongarch/loongarch.h (LARCH_CALL_RATIO):
	Modify the value of macro LARCH_CALL_RATIO on LoongArch to make it
        perform optimally.