Fix expected test output on hppa
Recent changes in the hoisting code change the optimized gimple for the
shadd-3 testcase on the PA.  That in turn changes the number of expected
shadd instructions.

I'm not entirely sure the test is actually testing what we want anymore
since I don't see a CSE for postreload to discover.  But I did verify
that the number of shadd instructions is sane, so I just changed the
count in the obvious way.

Pushed to the trunk.

Jeff
commit c0b88e9e8bbe15f0c2167371b49521b748c6da19
Author: Jeff Law <jlaw@ventanamicro.com>
Date:   Wed Jun 7 07:55:32 2023 -0600

    Fix expected test output on hppa

    Recent changes in the hoisting code change the optimized gimple for the
    shadd-3 testcase on the PA.  That in turn changes the number of expected
     shadd instructions.

    I'm not entirely sure the test is actually testing what we want anymore
    since I don't see a CSE for postreload to discover.  But I did verify
    that the number of shadd instructions is sane, so I just changed the
    count in the obvious way.

    gcc/testsuite
            * gcc.target/hppa/shadd-3.c: Update expected output.