c++: diagnostic ICE b/c of empty TPARMS_PRIMARY_TEMPLATE [PR109655]
When defining a previously declared class template, we neglect to set
TPARMS_PRIMARY_TEMPLATE for the in-scope template parameters, which the
class members go on to inherit, and so the members' DECL_TEMPLATE_PARMS
will have empty TPARMS_PRIMARY_TEMPLATE at those levels as well.  This
causes us to crash when diagnosing a constraint mismatch for an
out-of-line declaration of a member of a constrained class template.

This patch fixes this by walking the context to get at the corresponding
primary template instead.  I spent a while trying to get us to set
TPARMS_PRIMARY_TEMPLATE for templated class definitions that are
redeclarations, but it proved to be hairy in particular for partial
specializations and nested templates.

Bootstrapped and regtested on x86_64-pc-linux-gnu, does this look OK for
trunk/13?

	PR c++/109655

gcc/cp/ChangeLog:

	* pt.cc (push_template_decl): Handle TPARMS_PRIMARY_TEMPLATE
	being empty when diagnosing a constraint mismatch for an
	enclosing template scope.  Don't bother checking constraints
	if DECL_PARMS and SCOPE_PARMS are the same.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp2a/concepts-class6.C: New test.
	* g++.dg/cpp2a/concepts-class6a.C: New test.